chore: fix renamed files
