+++ apps/studio/src/components/CmsSidebar/index.tsx
@@ -1,3 +1,3 @@
 export * from "./CmsContainer"
 export * from "./CmsSidebar"
-export * from "./CmsSidebarOnlyContainer"
+export * from "./CmsSidebarContainer"
